ai-service-python/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
google-generativeai
python-dotenv
requests
scikit-learn
beautifulsoup4
